src/test/java/org/example/calculate/CalculatorTest.java
package org.example.calculate;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

/**
 * < 요구사항 > </>
 * 간단한 사칙연산을 할 수 있다.
 * 양수로만 계산할 수 있다.
 * 나눗셈에서 0을 나누는 경우 IllegalArgument 예외를 발생시킨다.
 * MVC패턴을 기반으로 구현하다.
 */
public class CalculatorTest {

    // 덧셈연산을 시행하기 위해서는 피연산자, 연산자, 피연산자를 calculator에 전달하면
    // calculator는 결과값을 다시 반환하면 된다.
    @DisplayName("덧셈 연산을 수행한다")
    @ParameterizedTest
    @MethodSource("formulaAndResult")
    void calculatorTest(int operand1, String operator, int operand2, int result) {
        int caculateResult = Calculator.calculate(new PositiveNumber(operand1), operator, new PositiveNumber(operand2));

        assertThat(caculateResult).isEqualTo(result);
    }

    private static Stream<Arguments> formulaAndResult() {
        return Stream.of(
                arguments(1, "+", 2, 3),
                arguments(5, "-", 2, 3),
                arguments(4, "*", 2, 8),
                arguments(4, "/", 2, 2)
        );
    }

//    @DisplayName("나눗셈에서 0을 나누는 경우 IllegalArgument 예외를 발생시킨다")
//    @Test
//    void calculateExceptionTest() {
//        assertThatCode(()-> Calculator.calculate(10, "/", 0))
//                .isInstanceOf(IllegalArgumentException.class);
//    }
}
